ReadMe
Application was developed as an assignment for the course Software Engineering (Programsko inzenjerstvo). The app is basically a simple “web streaming service” that allows users to stream videos from a remote vlc server. It is using a remote mysql database to store user data (profiles,roles,memberships) and allows them to search and watch movies or the add new movies to the program if they are administrators.